Adds a separated comment partial.

+++ public/js/post/comment.js
@@ -2,48 +2,55 @@ document.addEventListener('DOMContentLoaded', () => {
 
     // Initialize a new editor for each textarea.  
     document.querySelectorAll('[id^="tiny-comment-"]').forEach(function(item) { 
-        _initTinyMceEditor(item.dataset.commentId);
+        initTinyMceEditor(item.dataset.commentId);
     });
 
-    // Show the editor and hide the readonly comment.
-    document.querySelectorAll('[id^="edit-btn-"]').forEach(button => button.onclick = function() { 
-        //alert(this.id+' '+this.dataset.commentId);
-        document.getElementById('updateComment-'+this.dataset.commentId).style.display = 'block'; 
-        document.getElementById('comment-'+this.dataset.commentId).style.display = 'none'; 
+    document.getElementById('create-btn').onclick = function() { 
+        runAjax(document.getElementById('createComment'));
+    }
+
+    document.addEventListener('click', function(e) {
+        const deleteBtn = e.target.closest('[id^="delete-btn-"]');
+        const editBtn = e.target.closest('[id^="edit-btn-"]');
+        const updateBtn = e.target.closest('[id^="update-btn-"]');
+        const cancelBtn = e.target.closest('[id^="cancel-btn-"]');
+
+        // Delete the given comment.
+        if (deleteBtn) {
+            if (window.confirm('Are you sure ?')) {
+                runAjax(document.getElementById('deleteComment-'+deleteBtn.dataset.commentId));
+            }
+        }
+
+        if (editBtn) {
+            toggleEditor(editBtn.dataset.commentId);
+        }
+
+        if (updateBtn) {
+            runAjax(document.getElementById('updateComment-'+updateBtn.dataset.commentId));
+        }
+
+        if (cancelBtn) {
+            deleteMessages();
+            // Reset the editor to the original content.
+            tinyMCE.get('tiny-comment-'+cancelBtn.dataset.commentId).setContent(document.getElementById('comment-'+cancelBtn.dataset.commentId).innerHTML);
+            toggleEditor(cancelBtn.dataset.commentId);
+        }
     });
 
-    // Hide the editor and show the readonly comment.
-    document.querySelectorAll('[id^="cancel-btn-"]').forEach(button => button.onclick = function() { 
-        deleteMessages();
-        // Reset the editor to the original content.
-        tinyMCE.get('tiny-comment-'+this.dataset.commentId).setContent(document.getElementById('comment-'+this.dataset.commentId).innerHTML);
-        document.getElementById('updateComment-'+this.dataset.commentId).style.display = 'none'; 
-        document.getElementById('comment-'+this.dataset.commentId).style.display = 'block'; 
-    });
-
-    // Update the given comment.
-    document.querySelectorAll('[id^="update-btn-"]').forEach(button => button.onclick = function() { 
-        //alert(tinyMCE.get('comment-'+this.dataset.commentId).getContent());
-            //console.log(tinyMCE.get('comment-'+this.dataset.commentId).getContent());
-
-        let formData = new FormData(document.getElementById('updateComment-'+this.dataset.commentId));
+    function runAjax(form) {
+        let formData = new FormData(form);
 
         let ajax = new C_Ajax.init({
             method: 'post',
-            url: document.getElementById('updateComment-'+this.dataset.commentId).action,
+            url: form.action,
             dataType: 'json',
             data: formData,
             headers: {'X-CSRF-TOKEN': document.querySelector('meta[name="csrf-token"]').getAttribute('content'), 'Accept': 'application/json'}
         });
 
         ajax.run(getAjaxResult);
-    });
-
-    // Delete the given comment.
-    document.querySelectorAll('[id^="delete-btn-"]').forEach(button => button.onclick = function() { 
-        alert(this.id+' '+this.dataset.commentId);
-        document.getElementById('deleteComment-'+this.dataset.commentId).submit();
-    });
+    }
 
     function getAjaxResult(status, result) {
         //const spinner = document.getElementById('ajax-progress');
@@ -51,15 +58,29 @@ document.addEventListener('DOMContentLoaded', () => {
 
         if (status === 200) {
             deleteMessages();
-            // Update the readonly comment.
-            document.getElementById('comment-'+result.id).innerHTML = tinyMCE.get('tiny-comment-'+result.id).getContent();
-            // Hide the editor and show the readonly comment.
-            document.getElementById('updateComment-'+result.id).style.display = 'none'; 
-            document.getElementById('comment-'+result.id).style.display = 'block'; 
-            displayMessage('success', result.message, result.id);
+
+            if (result.action == 'create') {
+                document.getElementById('createComment').insertAdjacentHTML('afterend', result.render);
+                initTinyMceEditor(result.id);
+                tinyMCE.get('tiny-comment-'+result.id).setContent(result.text);
+                tinyMCE.get('tiny-comment-0').setContent('');
+                displayMessage('success', result.message, result.id);
+            }
+
+            if (result.action == 'update') {
+                // Update the readonly comment.
+                document.getElementById('comment-'+result.id).innerHTML = tinyMCE.get('tiny-comment-'+result.id).getContent();
+                toggleEditor(result.id);
+                displayMessage('success', result.message, result.id);
+            }
+
+            if (result.action == 'delete') {
+                document.getElementById('card-comment-'+result.id).remove();
+                displayMessage('success', result.message, 0);
+            }
+
         }
         else if (status === 422) {
-      console.log(result);
             displayMessage('danger', result.message, result.commentId);
             // Loop through the returned errors and set the messages accordingly.
             for (const [name, message] of Object.entries(result.errors)) {
@@ -96,10 +117,10 @@ document.addEventListener('DOMContentLoaded', () => {
         messageAlert.classList.remove('d-none', 'alert-success', 'alert-danger', 'alert-warning', 'alert-info');
         messageAlert.classList.add('alert-'+type);
         document.getElementById('ajax-message-'+idNb).innerHTML = message;
-        document.getElementById('card-comment-'+idNb).scrollIntoView({behavior: 'smooth', block: 'start', inline: 'nearest'});
+        document.getElementById('ajax-message-alert-'+idNb).scrollIntoView({behavior: 'smooth', block: 'start', inline: 'nearest'});
     }
 
-    function _initTinyMceEditor(idNb) {
+    function initTinyMceEditor(idNb) {
         let editor = tinymce.init({
 	    selector: '#tiny-comment-'+idNb,
             entity_encoding: 'raw',
@@ -110,10 +131,21 @@ document.addEventListener('DOMContentLoaded', () => {
                 editor.on('change', function () {
                     editor.save();
                 });
-
             }
 	});
 
         return editor;
+    }
+
+    // Hide or show the given editor.
+    function toggleEditor(idNb) {
+        if (document.getElementById('updateComment-'+idNb).style.display == 'none') {
+            document.getElementById('updateComment-'+idNb).style.display = 'block'; 
+            document.getElementById('comment-'+idNb).style.display = 'none'; 
+        }
+        else {
+            document.getElementById('updateComment-'+idNb).style.display = 'none'; 
+            document.getElementById('comment-'+idNb).style.display = 'block'; 
+        }
     }
 });
